NYC_Taxi_Analytics_Pipeline/ingesting_ny_taxi_data/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-microsoft-azure
apache-airflow-providers-http
apache-airflow-providers-microsoft-mssql 
apache-airflow-providers-odbc
